Add Product Service and Controller, configure basic security filter chain

+++ src/main/java/com/pjo/InventoryManagementSystem/models/Product.java
@@ -29,10 +29,10 @@ public class Product {
     private String description;
 
     @Column(nullable = false)
-    private BigDecimal price;
+    private Double price;
 
     @Column(nullable = false)
-    private int stockQuantity;
+    private Integer stockQuantity;
 
     @Column(nullable = false, updatable = false)
     @Temporal(TemporalType.TIMESTAMP)
